Add new parameter "exit_method".
The exit method defines how the applicatrion terminates.
Up to now the application stopped the CPU with an endless loop containing a "wait for something" instruction, like a WFE. A NOP is used where this is not available. This is now available as the "Stop" exit method.
The new "Return" method just returns, which is usually the ROM code.
The "Reset" method is also new. It resets the board.