Prevent the optimizer from ignoring patched options.
The parameters for the standalone version are patched into the binary after
compiliation. A direct access to the structure with the default values lead the
optimizer to the assumption that the values were constant. The result was
that the default values were hard-coded and the patched values were ignored.
Now the parameter structure is passed as a pointer from the init assembly
code, which prevented the unwanted optimization.